Add weather icon and description, and refactor display functions to a module
