+++ src/index.js
@@ -1,29 +1,9 @@
 import formattedWeatherData from './weather-api';
+import { displayLoadingGif, displayWeather } from './display';
 
 const apiKey = '3f4751a98d27d8f410e99e9589eaf508';
 const content = document.getElementById('content');
 const form = document.forms[0];
-
-function displayWeather(data, content) {
-  const location = document.createElement('p');
-  location.textContent = `${data.city}, ${data.country}`;
-  const temp = document.createElement('p');
-  temp.textContent = `${data.temp}K`;
-  content.innerHTML = '';
-  content.appendChild(location);
-  content.appendChild(temp);
-}
-
-function displayLoadingGif(content) {
-  if (document.querySelector('#loading')) {
-    return null;
-  }
-  const loading = document.createElement('img');
-  loading.src = '../src/loading.gif';
-  loading.id = 'loading';
-  content.appendChild(loading);
-  return true;
-}
 
 form.addEventListener('submit', async e => {
   e.preventDefault();
